+++ examples/scripts/SPMe.py
@@ -8,7 +8,8 @@ import numpy as np
 pybamm.set_logging_level("INFO")
 
 # load model
-model = pybamm.lithium_ion.DFN({"surface form": "differential"})
+model = pybamm.lithium_ion.SPMe()
+model.convert_to_format = "python"
 
 # create geometry
 geometry = model.default_geometry