+++ pybamm/models/submodels/electrolyte/stefan_maxwell/conductivity/surface_potential_form/leading_surface_form_stefan_maxwell_conductivity.py
@@ -49,9 +49,9 @@ class BaseLeadingOrderSurfaceForm(LeadingOrder):
             + " electrode surface potential difference"
         ]
         if self.domain == "Negative":
-            delta_phi_init = self.param.U_n(self.param.c_n_init, self.param.T_init)
+            delta_phi_init = self.param.U_n(self.param.c_n_init(0), self.param.T_init)
         elif self.domain == "Positive":
-            delta_phi_init = self.param.U_p(self.param.c_p_init, self.param.T_init)
+            delta_phi_init = self.param.U_p(self.param.c_p_init(1), self.param.T_init)
 
         self.initial_conditions = {delta_phi: delta_phi_init}