+++ pybamm/models/submodels/external_circuit/function_control_external_circuit.py
@@ -8,9 +8,9 @@ from .base_external_circuit import BaseModel, LeadingOrderBaseModel
 class FunctionControl(BaseModel):
     """External circuit with an arbitrary function. """
 
-    def __init__(self, param, external_circuit_class):
+    def __init__(self, param, external_circuit_function):
         super().__init__(param)
-        self.external_circuit_class = external_circuit_class
+        self.external_circuit_function = external_circuit_function
 
     def _get_current_variable(self):
         return pybamm.Variable("Total current density")
@@ -22,14 +22,6 @@ class FunctionControl(BaseModel):
 
         # Add discharge capacity variable
         variables.update(super().get_fundamental_variables())
-
-        # Add switches
-        # These are not implemented yet but can be used later with the Experiment class
-        # to simulate different external circuit conditions sequentially within a
-        # single model (for example Constant Current - Constant Voltage)
-        # for i in range(self.external_circuit_class.num_switches):
-        #     s = pybamm.Parameter("Switch {}".format(i + 1))
-        #     variables["Switch {}".format(i + 1)] = s
 
         return variables
 
@@ -44,7 +36,7 @@ class FunctionControl(BaseModel):
         # The external circuit function should fix either the current, or the voltage,
         # or a combination (e.g. I*V for power control)
         i_cell = variables["Total current density"]
-        self.algebraic[i_cell] = self.external_circuit_class(variables)
+        self.algebraic[i_cell] = self.external_circuit_function(variables)
 
 
 class VoltageFunctionControl(FunctionControl):
@@ -53,31 +45,25 @@ class VoltageFunctionControl(FunctionControl):
     """
 
     def __init__(self, param):
-        super().__init__(param, ConstantVoltage())
+        super().__init__(param, constant_voltage)
 
 
-class ConstantVoltage:
-    num_switches = 0
-
-    def __call__(self, variables):
-        V = variables["Terminal voltage [V]"]
-        return V - pybamm.FunctionParameter("Voltage function [V]", pybamm.t)
+def constant_voltage(variables):
+    V = variables["Terminal voltage [V]"]
+    return V - pybamm.FunctionParameter("Voltage function [V]", pybamm.t)
 
 
 class PowerFunctionControl(FunctionControl):
     """External circuit with power control. """
 
     def __init__(self, param):
-        super().__init__(param, ConstantPower())
+        super().__init__(param, constant_power)
 
 
-class ConstantPower:
-    num_switches = 0
-
-    def __call__(self, variables):
-        I = variables["Current [A]"]
-        V = variables["Terminal voltage [V]"]
-        return I * V - pybamm.FunctionParameter("Power function [W]", pybamm.t)
+def constant_power(variables):
+    I = variables["Current [A]"]
+    V = variables["Terminal voltage [V]"]
+    return I * V - pybamm.FunctionParameter("Power function [W]", pybamm.t)
 
 
 class LeadingOrderFunctionControl(FunctionControl, LeadingOrderBaseModel):
@@ -97,12 +83,12 @@ class LeadingOrderVoltageFunctionControl(LeadingOrderFunctionControl):
     """
 
     def __init__(self, param):
-        super().__init__(param, ConstantVoltage())
+        super().__init__(param, constant_voltage)
 
 
 class LeadingOrderPowerFunctionControl(LeadingOrderFunctionControl):
     """External circuit with power control, at leading order. """
 
     def __init__(self, param):
-        super().__init__(param, ConstantPower())
+        super().__init__(param, constant_power)