+++ pybamm/models/submodels/particle/fickian/__init__.py
@@ -1,3 +1,2 @@
-from .base_fickian_particle import BaseModel
 from .fickian_many_particles import ManyParticles
 from .fickian_single_particle import SingleParticle