+++ pybamm/solvers/casadi_solver.py
@@ -94,7 +94,7 @@ class CasadiSolver(pybamm.BaseSolver):
             solution = self._run_integrator(integrator, y0_diff, y0_alg, inputs, t_eval)
             solution.termination = "final time"
             return solution
-        elif model.events == {}:
+        elif not model.events:
             pybamm.logger.info("No events found, running fast mode")
             integrator = self.get_integrator(model, t_eval, inputs)
             y0_diff, y0_alg = np.split(model.y0, [rhs_size])
@@ -104,11 +104,10 @@ class CasadiSolver(pybamm.BaseSolver):
         elif self.mode == "safe":
             # Step-and-check
             init_event_signs = np.sign(
-                np.concatenate([event(0, model.y0) for event in model.events_eval])
+                np.concatenate([event(0, model.y0)
+                                for event in model.terminate_events_eval])
             )
-            pybamm.logger.info(
-                "Start solving {} with {} in 'safe' mode".format(model.name, self.name)
-            )
+            pybamm.logger.info("Start solving {} with {}".format(model.name, self.name))
             t = t_eval[0]
             y0 = model.y0
             # Initialize solution
@@ -149,7 +148,7 @@ class CasadiSolver(pybamm.BaseSolver):
                     np.concatenate(
                         [
                             event(0, current_step_sol.y[:, -1])
-                            for event in model.events_eval
+                            for event in model.terminate_events_eval
                         ]
                     )
                 )