+++ pybamm/solvers/scikits_dae_solver.py
@@ -66,7 +66,7 @@ class ScikitsDaeSolver(pybamm.BaseSolver):
         """
         residuals = model.residuals_eval
         y0 = model.y0
-        events = model.events_eval
+        events = model.terminate_events_eval
         jacobian = model.jacobian_eval
         mass_matrix = model.mass_matrix.entries